Refactor sphinx portions into ZConfig.sphinx.
