+++ doc/conf.py
@@ -34,7 +34,7 @@ extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.intersphinx',
     'sphinxcontrib.programoutput',
-    'ZConfig.schema2html',
+    'ZConfig.sphinx',
 ]
 
 # Add any paths that contain templates here, relative to this directory.
